refacto GroupRoleInterface and update views
