+++ BackofficeBundle/Resources/public/coffee/groupTree/groupTreeLib.coffee
@@ -1,9 +1,9 @@
 (($, OpenOrchestra) ->
-  OpenOrchestra.toggleTreeNodeDisplay = (e, childrenClass) ->
+  OpenOrchestra.toggleTreeNodeDisplay = (e) ->
     switcher = $(e.currentTarget)
     switcher.toggleClass 'open'
     nodeElement = switcher.closest('li')
-    container = nodeElement.children(childrenClass).first()
+    container = nodeElement.children('.child-document').first()
     if !switcher.hasClass('open')
       switcher.addClass 'fa-plus-square-o'
       switcher.removeClass 'fa-minus-square-o'
