+++ BackofficeBundle/Resources/public/less/groupTreePanel.less
@@ -1,10 +1,10 @@
-#tab-node_tree {
+#tab-node_tree, #tab-media_folder_tree {
 
   .header-role-list {
     margin-bottom: 10px;
   }
 
-  .content-node-list {
+  .content-document-list {
     li div.row {
       padding-bottom: 5px;
 
@@ -12,33 +12,35 @@
         padding: 0 2px;
       }
     }
-    .list-page-node-name {
+    
+    .list-document-name {
       overflow: hidden;
       padding-top: 7px;
       text-overflow: ellipsis;
       white-space: nowrap;
     }
+
     // second level
-    .child-node {
-      .list-page-node-name {
+    .child-document {
+      .list-document-name {
         padding-left: 25px;
       }
 
       // Third Level
-      .child-node {
-        .list-page-node-name {
+      .child-document {
+        .list-document-name {
           padding-left: 50px;
         }
 
         // Fourth Level
-        .child-node {
-          .list-page-node-name {
+        .child-document {
+          .list-document-name {
             padding-left: 75px;
           }
 
           // Fifth Level
-          .child-node {
-            .list-page-node-name {
+          .child-document {
+            .list-document-name {
               padding-left: 100px;
             }
           }
