+++ BackofficeBundle/Resources/views/BackOffice/Underscore/groupTree/nodeTree._tpl.twig
@@ -3,5 +3,5 @@
     <div class="col-sm-9 head-element"></div>
 </div>
 
-<ul class="content-node-list" data-title="{{ 'open_orchestra_backoffice.form.group.node_role_edition'|trans }}">
+<ul class="content-document-list" data-title="{{ 'open_orchestra_backoffice.form.group.node_role_edition'|trans }}">
 </ul>
