+++ BackofficeBundle/Resources/views/BackOffice/Underscore/groupTree/nodeTreeElement._tpl.twig
@@ -1,6 +1,6 @@
 <li data-title="{{ 'open_orchestra_backoffice.form.group.node_role_edition'|trans }}">
     <div class="row">
-        <div class="col-sm-3 list-page-node-name">
+        <div class="col-sm-3 list-document-name">
             <% if ( nodes.children.length > 0 ){ %>
                 <em class="fa fa-minus-square-o open"></em>
             <% } %>
@@ -10,5 +10,5 @@
             <div class="row form-input"></div>
         </div>
     </div>
-    <ul class="child-node"></ul>
+    <ul class="child-document"></ul>
 </li>
